% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_color.R
\name{covdata_color}
\alias{covdata_color}
\title{covdata_color}
\usage{
covdata_color(...)
}
\arguments{
\item{...}{Accepts the color name in 'covington_colors'}
}
\description{
This function allows you to access the HEX color code
options in 'covington_colors' by name in applying color or fill ggplot aesthetics.
It is best to access this variable to apply color to the data
components of the graphics plot (i.e., points, bars, lines, etc.)
when only one color is needed.
}
\examples{
ggplot +
geom_bar(stat = "identity",
        fill = covdata_color("Covington.Blue"))
}
\keyword{covdata_color}
