% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_color_c.R
\name{covdata_color_c}
\alias{covdata_color_c}
\title{covdata_color_c}
\usage{
covdata_color_c(name, reverse = FALSE)
}
\arguments{
\item{name}{is the name of the color palette from 'covdata_palette_colors'.}

\item{reverse}{is used to change the default order of colors contained in the selected palette.}
}
\description{
This function is used to apply a color aesthetic to a continuous variable.
Color palette options come from 'covdata_palette_colors'.
}
\examples{
cov_graph +
geom_point(stat = "identity", aes(color = Amount), size = 4)+
covdata_color_c("YlOrRd")

}
\references{
Jumping Rivers--Custom colour palettes for {ggplot2}
}
\keyword{covdata_color_c}
