% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_color_d.R
\name{covdata_color_d}
\alias{covdata_color_d}
\title{covdata_color_d}
\usage{
covdata_color_d(name, reverse = FALSE)
}
\arguments{
\item{name}{is the name of the color palette from 'covdata_palette_colors'.}

\item{reverse}{is used to change the default order of colors contained in the selected palette.}
}
\description{
This function is used to apply a color aesthetic to a discrete variable.
Color palette options come from 'covdata_palette_colors'.
options in 'covington_colors' by name in applying color or fill ggplot aesthetics.
}
\examples{
cov_graph +
geom_point(stat = "identity", aes(color = Group), size = 4)+
covdata_color_d("Pantones", reverse = TRUE)

}
\keyword{covdata_color_d}
