% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_fill_d.R
\name{covdata_fill_d}
\alias{covdata_fill_d}
\title{covdata_color_d}
\usage{
covdata_fill_d(name, reverse = FALSE)
}
\arguments{
\item{name}{is the name of the color palette from 'covdata_palette_colors'.}

\item{reverse}{is used to change the default order of colors contained in the selected palette.}
}
\description{
This function is used to apply a fill aesthetic to a discrete variable.
Color palette options come from 'covdata_palette_colors'.
}
\examples{
cov_graph +
geom_bar(stat = "identity", aes(fill = Group))+
covdata_fill_d("Pantones")

}
\keyword{covdata_fill_d}
