% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_palettes.R
\name{covdata_palettes}
\alias{covdata_palettes}
\title{covdata_palettes}
\usage{
covdata_palettes(
  name,
  n,
  all_palettes = covdata_palette_colors,
  type = c("discrete", "continuous"),
  reverse = FALSE
)
}
\arguments{
\item{name}{is the name of the color palette in covdata_palette_colors.}

\item{n}{is the number of colors from the color palette to use.}

\item{all_palettes}{is the list of color palettes to use.  The default choice
is set to use 'covdata_palette_colors.}

\item{type}{determines a discrete or continuous color palette.}

\item{reverse}{is used to change the default order of colors in the selected color palette.}
}
\description{
This function is used to generate a color palette from the
list of palettes contained in 'covdata_palette_colors'.
}
\examples{
covdata_palettes("Brand")
}
\references{
Jumping Rivers--Custom colour palettes for {ggplot2}
}
\keyword{covdata_palettes}
