% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_palettes_all.R
\name{covdata_palettes_all}
\alias{covdata_palettes_all}
\title{covdata_palettes_all}
\usage{
covdata_palettes_all(palette_list, palette_names)
}
\arguments{
\item{palette_list}{is the list containing color palettes.}

\item{palette_names}{is the names of the color palettes.}
}
\description{
This function is used to visualize all palettes simultaneously from the list of color palettes
contained in'covdata_palette_colors'.  The function can also be used to visually display any list of color palettes.
}
\examples{
palette_list <- rev(covdata_palette_colors)
palette_names <- rev(names(covdata_palette_colors))
covdata_palettes_all(palette_list, palette_names)
}
\keyword{covdata_palettes_all}
