% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_palettes_print.R
\name{covdata_palettes_print}
\alias{covdata_palettes_print}
\title{covdata_palettes_print}
\usage{
covdata_palettes_print(name, n, ...)
}
\arguments{
\item{name}{is the name of the color palette in covdata_palette_colors.}

\item{n}{is the number of colors from the color palette to display.}

\item{...}{additional arguments from 'covdata_palettes' (e.g., type, reverse)}
}
\description{
This function is used to visualize a single color palette from the list of color palettes
contained in'covdata_palette_colors'
}
\examples{
covdata_palettes_print("Brand")
covdata_palettes_print("Pantones")
covdata_palettes_print("RdBu", n = 26, type = "continuous")
}
\keyword{covdata_palettes_print}
