% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_plot.R
\name{covdata_plot}
\alias{covdata_plot}
\title{covdata_plot}
\usage{
covdata_plot(
  plot,
  logo = c("cov.seal.blue", "cov.seal.white", "cov.logo.blue", "cov.logo.white",
    "cov.logo.pride"),
  save_name,
  save_width = 5,
  save_height = 4
)
}
\arguments{
\item{plot}{the ggplot containing the data and style components}

\item{logo}{the name of the City logo from 'covington_logos'.}

\item{save_name}{the name (including file path) to use to save the graphics plot.  The suffix '.png' is not
needed as this and the date (ymd) is added automatically to the name.}

\item{save_width}{the width in inches for the saved graphics plot.  The default is 5 'in'}

\item{save_heght}{the height in inches for the saved graphics plot.  The default is 4 'in'}
}
\description{
This function will create and save a graphics plot in the
style that is consistent with the professional brand for the City of Covington.
It will add a footer that includes the COV logo at the bottom left and a horizontal
line across the entire plot width in the color covington.blue
}
\details{
For a correctly formatted footer, the plot must contain a Title, Subtitle,
an X-Axis label and a three line caption.
}
\examples{
covdata_plot(plot = plot,
logo = "cov.logo.blue",
save_name = "plot_name")
}
\keyword{covdata_plot}
