% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_plot_any.R
\name{covdata_plot_any}
\alias{covdata_plot_any}
\title{covdata_plot_any}
\usage{
covdata_plot_any(
  plot,
  logo,
  logo_width = 1.4,
  save_name,
  save_width = 5,
  save_height = 4
)
}
\arguments{
\item{plot}{the ggplot containing the data and style components.}

\item{logo}{the name of the logo to use.}

\item{logo_width}{the width adjustment in inches to the logo (default is set to 1.4)}

\item{save_name}{the name (including file path) to use to save the graphics plot.  The suffix '.png' is not
needed as this and the date (ymd) is added automatically to the name.}

\item{save_width}{the width in inches for the saved graphics plot.  The default is 5 'in'}

\item{save_heght}{the height in inches for the saved graphics plot.  The default is 4 'in'}
}
\description{
This function will create and save a graphics plot in the
style that is consistent with the professional brand for the City of Covington.
It will add a footer that can include any logo at the bottom left and a horizontal
line across the entire plot width in the color covington.blue.
}
\details{
For a correctly formatted footer, the plot must contain a Title, Subtitle,
an X-Axis label and a three line caption.
}
\examples{
covdata_plot_any(plot = plot,
logo = "cov.logo.BLUEtm.png",
save_name = "plot_name")
}
\keyword{covdata_plot_any}
