% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covdata_theme.R
\name{covdata_theme}
\alias{covdata_theme}
\title{covdata_theme}
\usage{
covdata_theme(
  plot.background.fill = c("covington.plain", "covington.black", "covington.blue",
    "covington.navy", "covington.dark.blue", "covington.light.blue", "covington.green",
    "covington.yellow", "covington.gray", "covington.light.gray", "covington.og.blue"),
  panel.background.fill = c("covington.plain", "covington.light.gray", "covington.gray",
    "covington.black", "covington.blue", "covington.light.blue"),
  text.color = c("covington.black", "covington.plain")
)
}
\arguments{
\item{plot.background.fill}{changes the plot background color using options available in covington_colors.  The
default color is white (covington.plain).}

\item{panel.background.fill}{changes the panel background color using options available in covington_colors. The
default color is white (covington.plain).}

\item{text.color}{changes the text color of all text displayed in the graphics plot (except for data labels). The
default color is black (covington.black).}
}
\description{
This function is used to produce graphics that have a
style that is consistent with the professional brand for the City of Covington. The function
modifies the 'theme_bw' default style of the ggplot2 graphics library.
}
\examples{
covdata_theme()#Covington.Basic Style
covdata_theme(plot.background.fill = "covington.gray", text.color = "covington.plain")
}
\keyword{covdata_theme}
