% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covington_colors.R
\docType{data}
\name{covington_colors}
\alias{covington_colors}
\title{covington_colors}
\format{
An object of class \code{character} of length 12.
}
\usage{
covington_colors
}
\description{
covington_colors is a character string that contains the
Pantones and Neutral branding colors, and the original Covington
Blue (covington.og.blue).
}
\examples{
covington_colors <<- c("covington.blue" = "#0047ba",
                      "covington.navy" = "#211261",
                      "covington.dark.blue" = "#000b8c",
                      "covington.light.blue" = "#00c1de",
                      "covington.green" = "#38d430",
                      "covington.red" = "#ea0029",
                      "covington.yellow" = "#ffc600",
                      "covington.black" = "#000000",
                      "covington.gray" = "#878787",
                      "covington.light.gray" = "#cccccc",
                      "covington.plain" = "#ffffff",
                      "covington.og.blue" = "#46b5d2")
}
\keyword{covington_colors}
