% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covington_logos.R
\docType{data}
\name{covington_logos}
\alias{covington_logos}
\title{covington_logos}
\format{
An object of class \code{character} of length 5.
}
\usage{
covington_logos
}
\description{
covington_logos is a character string that contains the
names of the four covington logos that are permitted with the CovData graphics
style.  Each of the '.png' logo files must be available in the working directory
}
\examples{
covington_logos <<- c("cov.seal.blue"="cov.seal.BLUE.png",
                      "cov.seal.white"="cov.seal.WHITE.png",
                      "cov.logo.blue"="cov.logo.BLUEtm.png",
                      "cov.logo.white"="cov.logo.WHITEtm.png",
                      "cov.logo.pride"="cov.logo.pride.png")
}
\keyword{covington_logos}
