% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covington_colors.R
\docType{data}
\name{covington_colors}
\alias{covington_colors}
\title{covington_colors}
\format{
An object of class \code{character} of length 12.
}
\usage{
covington_colors
}
\description{
covington_colors is a character string that contains the
Pantones and Neutral branding colors, and the original Covington
Blue (Covington.OG.Blue).
}
\examples{
covington_colors <<- c("Covington.Blue" = "#0047ba",
                      "Covington.Navy" = "#211261",
                      "Covington.Dark.Blue" = "#000b8c",
                      "Covington.Light.Blue" = "#00c1de",
                      "Covington.Green" = "#38d430",
                      "Covington.Red" = "#ea0029",
                      "Covington.Yellow" = "#ffc600",
                      "Covington.Black" = "#000000",
                      "Covington.Gray" = "#878787",
                      "Covington.Light.Gray" = "#cccccc",
                      "Covington.Plain" = "#ffffff",
                      "Covington.OG.Blue" = "#46b5d2")
}
\keyword{covington_colors}
